--- aviral.py
@@ -1,9 +1,0 @@
-from gauri import Gauri
-
-
-class Aviral:
-    def go_to_newyork(self):
-        ob = Gauri()
-        if ob.available_in_new_york():
-            return True
-        return False
