--- bar.py
@@ -1,2 +1,0 @@
-def name1(dolor):
-    return "aviral from bar"
