--- classes_sample_for_eric_sir.dot.json
@@ -1,221 +1,0 @@
-{
-  "name": "classes_sample_for_eric_sir",
-  "directed": true,
-  "strict": false,
-  "_draw_": 
-  [
-    {
-      "op": "c",
-      "grad": "none",
-      "color": "#fffffe00"
-    },
-    {
-      "op": "C",
-      "grad": "none",
-      "color": "#ffffff"
-    },
-    {
-      "op": "P",
-      "points": [[0.000,0.000],[0.000,178.000],[54.000,178.000],[54.000,0.000]]
-    }
-  ],
-  "bb": "0,0,54,178",
-  "charset": "utf-8",
-  "rankdir": "BT",
-  "xdotversion": "1.7",
-  "_subgraph_cnt": 0,
-  "objects": [
-    {
-      "_gvid": 0,
-      "name": "0",
-      "_draw_": 
-      [
-        {
-          "op": "c",
-          "grad": "none",
-          "color": "#000000"
-        },
-        {
-          "op": "p",
-          "points": [[0.000,0.500],[0.000,70.500],[54.000,70.500],[54.000,0.500]]
-        },
-        {
-          "op": "c",
-          "grad": "none",
-          "color": "#000000"
-        },
-        {
-          "op": "L",
-          "points": [[0.000,47.500],[54.000,47.500]]
-        },
-        {
-          "op": "c",
-          "grad": "none",
-          "color": "#000000"
-        },
-        {
-          "op": "L",
-          "points": [[0.000,23.500],[54.000,23.500]]
-        }
-      ],
-      "_ldraw_": 
-      [
-        {
-          "op": "F",
-          "size": 14.000,
-          "face": "Times-Roman"
-        },
-        {
-          "op": "c",
-          "grad": "none",
-          "color": "#000000"
-        },
-        {
-          "op": "T",
-          "pt": [27.000,55.300],
-          "align": "c",
-          "width": 10.000,
-          "text": "A"
-        },
-        {
-          "op": "F",
-          "size": 14.000,
-          "face": "Times-Roman"
-        },
-        {
-          "op": "c",
-          "grad": "none",
-          "color": "#000000"
-        },
-        {
-          "op": "T",
-          "pt": [27.000,8.300],
-          "align": "c",
-          "width": 4.000,
-          "text": " "
-        }
-      ],
-      "height": "0.98611",
-      "label": "{A|\\l|}",
-      "pos": "27,35.5",
-      "rects": "0,47.5,54,70.5 0,23.5,54,47.5 0,0.5,54,23.5",
-      "shape": "record",
-      "width": "0.75"
-    },
-    {
-      "_gvid": 1,
-      "name": "1",
-      "_draw_": 
-      [
-        {
-          "op": "c",
-          "grad": "none",
-          "color": "#000000"
-        },
-        {
-          "op": "p",
-          "points": [[0.000,107.500],[0.000,177.500],[54.000,177.500],[54.000,107.500]]
-        },
-        {
-          "op": "c",
-          "grad": "none",
-          "color": "#000000"
-        },
-        {
-          "op": "L",
-          "points": [[0.000,154.500],[54.000,154.500]]
-        },
-        {
-          "op": "c",
-          "grad": "none",
-          "color": "#000000"
-        },
-        {
-          "op": "L",
-          "points": [[0.000,130.500],[54.000,130.500]]
-        }
-      ],
-      "_ldraw_": 
-      [
-        {
-          "op": "F",
-          "size": 14.000,
-          "face": "Times-Roman"
-        },
-        {
-          "op": "c",
-          "grad": "none",
-          "color": "#000000"
-        },
-        {
-          "op": "T",
-          "pt": [27.000,162.300],
-          "align": "c",
-          "width": 10.000,
-          "text": "B"
-        },
-        {
-          "op": "F",
-          "size": 14.000,
-          "face": "Times-Roman"
-        },
-        {
-          "op": "c",
-          "grad": "none",
-          "color": "#000000"
-        },
-        {
-          "op": "T",
-          "pt": [27.000,115.300],
-          "align": "c",
-          "width": 4.000,
-          "text": " "
-        }
-      ],
-      "height": "0.98611",
-      "label": "{B|\\l|}",
-      "pos": "27,142.5",
-      "rects": "0,154.5,54,177.5 0,130.5,54,154.5 0,107.5,54,130.5",
-      "shape": "record",
-      "width": "0.75"
-    }
-  ],
-  "edges": [
-    {
-      "_gvid": 0,
-      "tail": 0,
-      "head": 1,
-      "_draw_": 
-      [
-        {
-          "op": "c",
-          "grad": "none",
-          "color": "#000000"
-        },
-        {
-          "op": "b",
-          "points": [[27.000,70.540],[27.000,79.030],[27.000,88.240],[27.000,97.140]]
-        }
-      ],
-      "_hdraw_": 
-      [
-        {
-          "op": "S",
-          "style": "solid"
-        },
-        {
-          "op": "c",
-          "grad": "none",
-          "color": "#000000"
-        },
-        {
-          "op": "p",
-          "points": [[23.500,97.390],[27.000,107.390],[30.500,97.390]]
-        }
-      ],
-      "arrowhead": "empty",
-      "arrowtail": "none",
-      "pos": "e,27,107.39 27,70.536 27,79.034 27,88.241 27,97.141"
-    }
-  ]
-}
