--- commands.sh
@@ -1,2 +1,0 @@
-pyreverse -S sample_class_module.py --all-ancestors
-dot -Tpdf classes.dot -o output.pdf
