+++ dependency_collector.py
