--- driver.py
@@ -1,4 +1,0 @@
-from aviral import Aviral
-
-aviral_ = Aviral()
-print(aviral_.go_to_newyork())
