--- gauri.py
@@ -1,9 +1,0 @@
-class Gauri:
-    def gauri_in_new_york(self):
-        value = "tatti kha"  # pull from db
-        return value
-
-    def available_in_new_york(self):
-        if self.gauri_in_new_york():
-            return True
-        return False
