--- inheritance_and_dependencies/another_module.py
@@ -1,8 +1,0 @@
-class ABC:
-    def add(self, x,y):
-        return (x+y)
-    
-
-class DEF(ABC):
-    def sub(self, x,y):
-        return (x-y)
