--- requirements.txt
@@ -1,58 +1,0 @@
-anytree==2.6.0
-astroid==2.2.5
-attrs==19.1.0
-autopep8==1.4.4
-backcall==0.1.0
-bleach==3.1.0
-decorator==4.4.0
-defusedxml==0.6.0
-entrypoints==0.3
-ipykernel==5.1.2
-ipython==7.7.0
-ipython-genutils==0.2.0
-ipywidgets==7.5.1
-isort==4.3.21
-jedi==0.15.1
-Jinja2==2.10.1
-jsonschema==3.0.2
-jupyter==1.0.0
-jupyter-client==5.3.1
-jupyter-console==6.0.0
-jupyter-core==4.5.0
-lazy-object-proxy==1.4.1
-MarkupSafe==1.1.1
-mccabe==0.6.1
-mistune==0.8.4
-nbconvert==5.6.0
-nbformat==4.4.0
-notebook==6.0.0
-numpy==1.17.0
-pandas==0.25.0
-pandocfilters==1.4.2
-parso==0.5.1
-pexpect==4.7.0
-pickleshare==0.7.5
-pkg-resources==0.0.0
-prometheus-client==0.7.1
-prompt-toolkit==2.0.9
-ptyprocess==0.6.0
-pycodestyle==2.5.0
-Pygments==2.4.2
-pylint==2.3.1
-pyrsistent==0.15.4
-python-dateutil==2.8.0
-pytz==2019.2
-pyzmq==18.1.0
-qtconsole==4.5.3
-Send2Trash==1.5.0
-six==1.12.0
-terminado==0.8.2
-testpath==0.4.2
-tornado==6.0.3
-traitlets==4.3.2
-typed-ast==1.4.0
-wcwidth==0.1.7
-webencodings==0.5.1
-widgetsnbextension==3.5.1
-wrapt==1.11.2
-XlsxWriter==1.1.8
