+++ sample_class_module.py
