--- sample_for_eric_sir.py
@@ -1,6 +1,0 @@
-class B:
-    pass
-
-
-class A(B):
-    pass
