--- source.py
@@ -1,15 +1,0 @@
-from bar import name1 as namealias1
-from foo import bar as modalias1
-
-def loremipsum(dolor):
-    return namealias1(dolor)
-
-def sitamet():
-    from foo.bar import consectetur
-
-    modalias1 = 'something else'
-    consectetur(modalias1)
-
-class Adipiscing:
-    def elit_nam(self):
-        return modalias1.name2(self)
